// projects/customer360/src/app/home/customer/change-of-mandate/models/change-madate.model.ts
export interface MandateResponse {
  statusMessage: string;
  statusCode: string;
  successful: boolean;
  responseObject: ResponseObject;
}

export interface ResponseObject {
  mandates: IMandate[];
}

export interface IMandate {
  accountId: string;
  accountNumber: string;
  operationMode: string;
  customerId: string;
  cif: string;
  email: string;
  phone: string;
  signatoryType: string;
  isDeleted: boolean;
  customerName: string;
  signatoryName: string;
}

// Generated by https://quicktype.io

export interface ICreateChangeMandateTicketPayload {
  parentTaskId?: string | null;
  associatedId?: string | null;
  mandateType: string;
  SpecialMandateRequest: string;
  startDate: string;
  bankId: string;
  customerDetails: CustomerDetails;
  CustomerPresent: boolean;
  ViewProfileTicketId?: string | null;
}

export interface CustomerDetails {
  accountId?: string;
  firstName?: string;
  middleName?: string;
  lastName?: string;
  customerId?: string;
  phone?: Phone;
  email?: Email;
}

export interface Email {
  id: string;
  emailAddress: string;
  emailType: string;
  preferred: boolean;
}

export interface Phone {
  id: string;
  code: string;
  number: string;
  phoneType: string;
  preferred: boolean;
}

// Generated by https://quicktype.io

export interface ITikectChangeMandateResponse {
  statusMessage: string;
  statusCode: string;
  successful: boolean;
  responseObject: ResponseObject;
}

export interface ResponseObject {
  taskTypeId: number;
  taskType: ActionFlow;
  actionFlowId: number;
  actionFlow: ActionFlow;
  taskData: string;
  status: string;
  taskDocuments: any[];
  taskNotes: any[];
  actions: Action[];
  createdBy: string;
  associatedId: string;
  userId: string;
  branchId: string;
  bankId: string;
  workItem: any;
  customerId: string;
  customerName: string;
  id: string;
  createdOnUtc: string;
  modifiedOnUtc: string;
  isDeleted: boolean;
}

export interface ActionFlow {
  name: string;
  description: any;
  branchId?: any;
  isSequential?: boolean;
  id: number;
  createdOnUtc: string;
  modifiedOnUtc: string;
  isDeleted: boolean;
}

export interface Action {
  taskId: number;
  actionTypeId: number;
  actionType: ActionFlow;
  status: string;
  order: number;
  error: null;
  isRepeatable: boolean;
  id: number;
  createdOnUtc: string;
  modifiedOnUtc: string;
  isDeleted: boolean;
}